% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aggregation}
\alias{aggregation}
\title{"Aggregation" data}
\format{
A data frame with 788 rows and 4 columns:
\describe{
\item{x, y}{Spatial coordinates of the original dataset}
\item{age}{Randomly sampled age}
\item{clust}{Simulated spatial cluster}
}
}
\source{
\url{https://cs.joensuu.fi/sipu/datasets/}
}
\usage{
aggregation
}
\description{
"Aggregation" data
}
\references{
Gionis, A., Mannila, H., & Tsaparas, P. (2007). Clustering aggregation. Acm transactions on knowledge discovery from data (tkdd), 1(1), 4-es.

Rezaei, M., & Fränti, P. (2020). Can the number of clusters be determined by external indices?. IEEE Access, 8, 89239-89257.
}
\keyword{datasets}
