% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{append_moo_out}
\alias{append_moo_out}
\title{Append MOO output}
\usage{
append_moo_out(pop, obj, pareto_idx, moo_quality)
}
\arguments{
\item{pop}{a list of objects returned from \code{\link[=blob_populate]{blob_populate()}} or \code{\link[=blob_kpopulate]{blob_kpopulate()}} from the current batch.}

\item{obj}{a string vector of objectives.}

\item{pareto_idx}{a numeric vector of indices of Pareto optimal solutions}

\item{moo_quality}{a data frame of MOO quality indicators.}
}
\value{
a list of the following objects.
\itemize{
\item \code{clust}: a numeric matrix of cluster assignments. Each row is a Pareto optimal solution.
\item \code{summary}: a data frame of summary statistics of all feasible solutions.
\item \code{trace}: a data frame of summary statistics for tracing of all feasible solutions.
\item \code{n_filtered}: a data frame of numbers of filtered solutions.
\item \code{clust_idx}: a numeric vector of indices of Pareto optimal cluster assignment.
\item \code{moo_quality}: a data frame of MOO quality indicators.
\item \code{obj}: a string vector of objectives.
}
}
\description{
This function appends all multi-objective optimsation (MOO) related output.
}
