% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asymmetric}
\alias{asymmetric}
\title{"Asymmetric" data}
\format{
A data frame with 1000 rows and 4 columns:
\describe{
\item{x, y}{Spatial coordinates of the original dataset}
\item{age}{Randomly sampled age}
\item{clust}{True spatial cluster}
}
}
\source{
\url{https://cs.joensuu.fi/sipu/datasets/}
}
\usage{
asymmetric
}
\description{
"Asymmetric" data
}
\references{
Fränti, P., & Sieranoja, S. (2018). K-means properties on six clustering benchmark datasets. Applied intelligence, 48(12), 4743-4759.

Rezaei, M., & Fränti, P. (2020). Can the number of clusters be determined by external indices?. IEEE Access, 8, 89239-89257.
}
\keyword{datasets}
