% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moo.R
\name{blob_populate_batch}
\alias{blob_populate_batch}
\title{Populate solutions by weighted sum scalarisation for multiple batches}
\usage{
blob_populate_batch(
  data,
  k,
  r = c(0.5, 1),
  iter = 10,
  run = 100,
  batch = 5,
  converge_ari = NULL,
  coords = c(1, 2),
  age = 3,
  crs = 4326,
  hull_convex_ratio = 0,
  random_start = FALSE,
  filter_intersects = TRUE,
  filter_clustsize = TRUE,
  max_na = 0.05,
  space_kmat = NULL,
  space_distmat = NULL,
  space_distmethod = NULL,
  w_knn = NULL,
  l_normalise = NULL,
  beta_par = NULL
)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer value or vector of length 2. If a vector is supplied, they specify the lower and upper bounds of the number of clusters.}

\item{r}{a numeric value or vector of length 2. If a vector is supplied, they specify the lower and upper bounds of the relative spatial weight.
They must be \eqn{[0,1]}. Default is c(0.5,1).}

\item{iter}{an integer of the number of iterations. Default is 10.}

\item{run}{an integer of the number of runs. Default is 100.}

\item{batch}{an integer of the number of batches. Default is 5.}

\item{converge_ari}{a numeric value of the Adjusted Rand Index (ARI) that sets the convergence threshold between iterations. It must be \eqn{[0,1]}. Default is NULL.}

\item{coords}{a vector of strings or numeric values indicating the columns of coordinates (longitude, latitide). Default is the first two columns.}

\item{age}{a string or numeric value indicating the column of age. Default is the third column.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}. Default is 4326.}

\item{hull_convex_ratio}{a numeric value controlling the convexity of the hulls passed onto \code{\link[sf:geos_unary]{sf::st_concave_hull()}}. 1 returns convex and 0 maximally concave hulls. Default is 0.}

\item{random_start}{a Boolean to randomise starting cluster members. Default is FALSE.}

\item{filter_intersects}{a Boolean to remove an assignment with intersects in space? Default is TRUE.}

\item{filter_clustsize}{a Boolean to assign NA to clusters below the critical size. Default is TRUE.}

\item{max_na}{a numeric value of the maximum proportion of NAs allowed. It must be \eqn{[0,1]}. Default is 0.05.}

\item{space_kmat}{a kernel matrix computed from the spatial distance matrix.}

\item{space_distmat}{a spatial distance matrix, used when \code{space_kmat} is not supplied. Default is NULL.}

\item{space_distmethod}{a string of the method, used when \code{space_kmat} and \code{space_distmat} are not specified. It must be one of "geodesic" or "euclidean". Default is NULL.}

\item{w_knn}{an integer of the k-th nearest neighbour, used in local scaling for computing the adjacency matrix with \code{\link[=compute_adjacency]{compute_adjacency()}} when \code{space_kmat} is not supplied. Default is 7.}

\item{l_normalise}{a Boolean to normalise the graph Laplacian, passed onto \code{\link[=compute_laplacian]{compute_laplacian()}} when \code{space_kmat} is not supplied. Default is TRUE.}

\item{beta_par}{an numeric initial value for the parameter \eqn{\beta} to be optimised over for the diffusion kernel with \code{\link[=optim_beta]{optim_beta()}} when \code{space_kmat} is not supplied. The range is \eqn{[0.001, \inf]}. Default is 10.}
}
\value{
a list of \code{pop} objects; see also \code{\link[=blob_populate]{blob_populate()}}.
}
\description{
This function populates solutions by weighted sum scalarisation of the bi-objective
search algorithm in \code{\link[=blob_search]{blob_search()}} for a given k for multiple batches.
}
\details{
A diffusion kernel is applied to compute the distance to the centroid in space.
See \code{\link[=distmat_to_kmat]{distmat_to_kmat()}} for more details for converting a distance matrix to a kernel matrix.

Clusters are assigned in every iteration. It iterates until the set length or convergence.

When \code{converge_ari} is specified, convergence is defined and activated when ARI between the latest and the previous search is
above the specified threshold and at least three iterations are run.

The critical size of a cluster is defined as \eqn{\frac{n}{2k}} where \eqn{n} is the number of data point and \eqn{k} is the number of clusters.

Scalarisation is achieved by varying the relative spatial weight generated by Latin hypercube sampling using \code{\link[lhs:randomLHS]{lhs::randomLHS()}}.

To parallelise runs, \code{\link[future.apply:future_lapply]{future.apply::future_lapply()}} is implemented. See \link[future:future]{future::future} and \link[future.apply:future.apply]{future.apply::future.apply} for more information.

\code{k} is set to \eqn{[2, \lfloor\sqrt{N}\rfloor]} when NULL.
}
\seealso{
\code{\link[=distmat_to_kmat]{distmat_to_kmat()}}, \code{\link[sf:st_as_sf]{sf::st_as_sf()}}, \code{\link[lhs:randomLHS]{lhs::randomLHS()}}, \code{\link[mclust:adjustedRandIndex]{mclust::adjustedRandIndex()}}, \link[future:future]{future::future}, \link[future.apply:future.apply]{future.apply::future.apply}
}
