% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{blob_search}
\alias{blob_search}
\title{Core local-search algorithm}
\usage{
blob_search(
  data,
  k,
  r,
  iter = 3L,
  space_distmat,
  sigma,
  converge_ari = NULL,
  crs = 4326,
  filter_intersects = T,
  filter_clustsize = T,
  max_na = 0.05,
  ...
)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters.}

\item{r}{an numeric value of spatial relative weight. It must be \eqn{[0,1]}.}

\item{iter}{an integer of the number of iterations. Default is 3L.}

\item{space_distmat}{a numeric spatial distance matrix.}

\item{sigma}{a numeric value of sigma passed on to \code{\link[=rbf]{rbf()}}.}

\item{converge_ari}{a numeric value of Adjusted Rand Index (ARI) that sets convergence threshold between two searches. It must be \eqn{[0,1]}. Default is NULL.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}} for geometry. Default is 4326.}

\item{filter_intersects}{a logical operator. Should an assignment with intersects in space be removed? Default is T.}

\item{filter_clustsize}{a logical operator. Should a cluster below the critical size be assigned NA? Default is T.}

\item{max_na}{a numeric value of the maximum proportion of NAs allowed. It must be \eqn{[0,1]}. Default is 0.05.}

\item{...}{the optional arguments include \code{random_start}.
\itemize{
\item \code{random_start}: a logical operator. Should random start or \code{\link[=start_blobs]{start_blobs()}} be used? Default is F.
}}
}
\value{
a list of the following objects,
a numeric value of 1 indicating being removed due to intersects,
2 indicating being removed due to the proportion of NAs exceeding \code{max_na}
or 3 indicating being removed due to \code{k == 1}.
\itemize{
\item \code{data}: a data frame of the input data with assigned clusters as a column.
\item \code{summary}: a data frame of summary statistics.
\item \code{clust_below_size}: a numeric vector of clusters below the critical size.
\item \code{trace}: a data frame of summary statistics for tracing.
}
}
\description{
This function performs the core bi-objective optimisation algorithm to assign clusters for a given k and r in a iterative fashion and
returns a feasible solution under the constraints.
}
\details{
Gaussian kernel is applied to compute cluster centroids and non-spherical clusters in space.

Clusters are assigned in every iteration. It iterates until the set length or convergence.

When \code{converge_ari} is specified, convergence is defined and activated when ARI between the latest and the previous search is
above the specified threshold and at least three iterations are run.

The critical size of a cluster is defined as \eqn{\frac{N}{2k}} where \eqn{N} is the number of data point and \eqn{k} is the number of clusters.
}
\seealso{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}}
}
