% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{blob_search_common}
\alias{blob_search_common}
\title{Core local-search algorithm - "commonality" method for convergence}
\usage{
blob_search_common(
  data,
  k,
  r,
  iter = 3L,
  space_distmat = NULL,
  space_distmethod = "geodesic",
  converge_ari = NULL,
  random_start = F,
  crs = 4326,
  ...
)
}
\arguments{
\item{data}{a data matrix or data frame with assigned (starting) clusters as a column.}

\item{k}{an integer of the number of clusters.}

\item{r}{an numeric value of spatial relative weight. It must be \eqn{[0,1]}.}

\item{iter}{an integer of the number of iterations. Default is 3L.}

\item{space_distmat}{a numeric spatial distance matrix. Default is NULL.}

\item{space_distmethod}{a string of the method for distance computation. It must be either "geodesic" or "euclidean". Default is "geodesic".}

\item{converge_ari}{a numeric value of Adjusted Rand Index (ARI) that sets convergence threshold between two searches. It must be \eqn{[0,1]}. Default is NULL.}

\item{random_start}{a logical operator. Should random start or \code{\link[=start_blobs]{start_blobs()}} be used? Default is F.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}. Default is 4326.}

\item{...}{these arguments include \code{sigma}.}
}
\value{
a list of the following objects.
\itemize{
\item \code{data}: a data frame of the input data with assigned clusters as a column.
\item \code{summary}: a data frame of summary statistics.
\item \code{clust_below_size}: a numeric vector of clusters below the critical size.
\item \code{trace}: a data frame of summary statistics for tracing.
}
}
\description{
This function performs the core bi-objective optimisation algorithm to assign clusters for a given k and r.
Convergence is directed by the "commonality" method.
}
\details{
When distances are computed on the fly, kernel method cannot be applied to compute cluster centroids in space.

When \code{space_distmat} is specified. Distance matrix implementation will ignore \code{space_distmethod}.
In that case, kernel method is applied to compute cluster centroids in space.

When the "commonality" method is specified, in every iteration, three searches are run and the shared members are retained for the next iteration.
After the last iteration or when searches converge, the search with the most shared members will be used to assign clusters.
When \code{converge_ari} is specified, convergence is defined and activated when all ARIs between all two pairs of searches is
above the specified threshold and at least three iterations are performed.
}
