% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{blob_search_filter}
\alias{blob_search_filter}
\title{Filter constraints from a local search}
\usage{
blob_search_filter(
  blob,
  k_matrix,
  crs,
  filter_intersects = T,
  filter_clustsize = T,
  max_na = 0.05
)
}
\arguments{
\item{blob}{a list of objects returned from \code{\link[=blob_search]{blob_search()}}.}

\item{k_matrix}{a kernel matrix computed from the spatial distance matrix.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}} for geometry. Default is 4326.}

\item{filter_intersects}{a logical operator. Should an assignment with intersects in space be removed? Default is T.}

\item{filter_clustsize}{a logical operator. Should a cluster below the critical size be assigned NA? Default is T.}

\item{max_na}{a numeric value of the maximum proportion of NAs allowed. It must be \eqn{[0,1]}. Default is 0.05.}
}
\value{
a list of the following objects,
a numeric value of 1 indicating being removed due to intersects,
2 indicating being removed due to the proportion of NAs exceeding \code{max_na}
or 3 indicating being removed due to \code{k == 1}.
\itemize{
\item \code{data}: a data frame of the input data with assigned clusters as a column.
\item \code{summary}: a data frame of summary statistics.
\item \code{clust_below_size}: a numeric vector of clusters below the critical size.
\item \code{trace}: a data frame of summary statistics for tracing.
}
}
\description{
This function filters data points or the entire solution when constraints are violated.
}
\seealso{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}}
}
