% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{blob_search_iter}
\alias{blob_search_iter}
\title{Core local-search algorithm}
\usage{
blob_search_iter(
  data,
  k,
  r,
  iter = 3L,
  space_distmat,
  sigma,
  converge_ari = NULL,
  crs = 4326,
  ...
)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters.}

\item{r}{an numeric value of spatial relative weight. It must be \eqn{[0,1]}.}

\item{iter}{an integer of the number of iterations. Default is 3L.}

\item{space_distmat}{a numeric spatial distance matrix.}

\item{sigma}{a numeric value of sigma passed on to \code{\link[=rbf]{rbf()}}.}

\item{converge_ari}{a numeric value of Adjusted Rand Index (ARI) that sets convergence threshold between two searches. It must be \eqn{[0,1]}. Default is NULL.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}} for geometry. Default is 4326.}

\item{...}{the optional arguments include \code{random_start}.
\itemize{
\item \code{random_start}: a logical operator. Should random start or \code{\link[=start_blobs]{start_blobs()}} be used? Default is F.
}}
}
\value{
a list of the following objects.
\itemize{
\item \code{data}: a data frame of the input data with assigned clusters as a column.
\item \code{summary}: a data frame of summary statistics.
\item \code{clust_below_size}: a numeric vector of clusters below the critical size.
\item \code{trace}: a data frame of summary statistics for tracing.
}
}
\description{
This function performs the core bi-objective optimisation algorithm to assign clusters for a given k and r in a iterative fashion.
}
\details{
Gaussian kernel is applied to compute cluster centroids and non-spherical clusters in space.

Clusters are assigned in every iteration. It iterates until the set length or convergence.

When \code{converge_ari} is specified, convergence is defined and activated when ARI between the latest and the previous search is
above the specified threshold and at least three iterations are run.
}
\seealso{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}}
}
