% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moo.R
\name{combine_pop}
\alias{combine_pop}
\title{Combine pop objects}
\usage{
combine_pop(pop_a, pop_b)
}
\arguments{
\item{pop_a, pop_b}{a \code{pop}, \code{pop_pareto} or \code{pop_moo} object;
see also \code{\link[=blob_populate]{blob_populate()}}, \code{\link[=find_pareto]{find_pareto()}} and \code{\link[=eval_moo]{eval_moo()}}.}
}
\value{
a \code{pop} object.
}
\description{
This function combines \code{pop} objects.
}
