% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{compare_blobs}
\alias{compare_blobs}
\title{Compare two cluster assignments}
\usage{
compare_blobs(b1, b2)
}
\arguments{
\item{b1}{a data frame with assigned clusters as a column.}

\item{b2}{a data frame with assigned clusters as a column.}
}
\value{
a data frame with shared assigned clusters as a column.
}
\description{
This function compares two cluster assignments and retains the shared members.
}
