% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{compute_adjacency}
\alias{compute_adjacency}
\title{Compute adjacency matrix using local scaling}
\usage{
compute_adjacency(distmat, knn = 7)
}
\arguments{
\item{distmat}{a distance matrix.}

\item{knn}{an integer of the k-th nearest neighbour. Default is 7.}
}
\value{
a weighted adjacency matrix.
}
\description{
This function converts a distance matrix to a weighted adjacency matrix using a local scaling similarity function.
}
\details{
The similarity function was proposed by Zelnik-Manor and Perona (2004).
Instead of using a fixed scaling parameter for all data points, a local scaling parameter is calculated for
every point based on their \eqn{k^{th}} nearest neighbour in the conversion of pairwise distance to pairwise adjacency (affinity/similarity).
See Zelnik-Manor and Perona (2004) for more information.
}
\references{
Zelnik-Manor, L., & Perona, P. (2004). Self-Tuning Spectral Clustering. Neural Information Processing Systems, 17, 1601–1608.
}
