% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{compute_dist}
\alias{compute_dist}
\title{Distance computation}
\usage{
compute_dist(x, y, ax, ay, method = "geodesic")
}
\arguments{
\item{x}{a numeric value of longitude.}

\item{y}{a numeric value of latitude.}

\item{ax}{a numeric vector of longitudes.}

\item{ay}{a numeric vector latitudes.}

\item{method}{a string of method, This must be one of "geodesic" or "euclidean". Default is "geodesic".}
}
\value{
a numeric vector of distances. Radian is the unit when "geodesic" is specified.
}
\description{
This function computes and returns either geodesic or Euclidean distances.
}
