% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{compute_distmat}
\alias{compute_distmat}
\title{Distance matrix computation}
\usage{
compute_distmat(data, method = "geodesic")
}
\arguments{
\item{data}{a numeric matrix or data frame of two columns. The first and second columns are longitude and latitude respectively.}

\item{method}{a string of method. It must be one of "geodesic" or "euclidean". Default is "geodesic".}
}
\value{
a numeric distance matrix. Km is the unit when "geodesic" is specified.
}
\description{
This function computes and returns the distance matrix computed using the specified method
to compute the distances between the rows of a data matrix. Km is the unit when "geodesic" is specified.
}
