% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{compute_erank}
\alias{compute_erank}
\title{Compute the effective rank of the diffusion kernel matrix}
\usage{
compute_erank(L, beta)
}
\arguments{
\item{L}{a graph Laplacian.}

\item{beta}{a numeric parameter controlling the diffusion rate.}
}
\value{
a numeric value of the effective rank.
}
\description{
This function computes and returns the effective rank of the diffusion kerenl matrix following Roy and Vetterli (2007).
}
\details{
As the diffusion kernel is symmetric and positive definite,
its eigenvalues \eqn{e^{\beta\lambda_i}}
are essentially singular values \eqn{\sigma_i}, where \eqn{\lamba_i} are eigenvalues of \eqn{L}.

Instead of treating \eqn{0\log{0} = 0}
when any singular values \eqn{\sigma_i = 0} in Roy and Vetterli (2007),
1e8 is returned for its application in the optimisation of the diffusion rate (\eqn{\beta}) for kernel k-means clustering
to constrain \eqn{\sigma_i = 0} > 0.
}
\references{
Kondor, R. I., & Lafferty, J. (2002, July). Diffusion kernels on graphs and other discrete structures. In Proceedings of the 19th international conference on machine learning (Vol. 2002, pp. 315-322).

Kondor, R., & Vert, J. P. (2004). Diffusion kernels. kernel methods in computational biology, 171-192.

Roy, O., & Vetterli, M. (2007, September). The effective rank: A measure of effective dimensionality. In 2007 15th European signal processing conference (pp. 606-610). IEEE.
}
