% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{compute_erank_sigma}
\alias{compute_erank_sigma}
\title{Compute the effective rank from singular values of a matrix}
\usage{
compute_erank_sigma(sigma)
}
\arguments{
\item{sigma}{a numeric vector of singular values.}
}
\value{
a numeric value of the effective rank.
}
\description{
This function computes and returns the effective rank from singular values of a matrix following Roy and Vetterli (2007).
}
\details{
Instead of treating \eqn{0\log{0} = 0} in Roy and Vetterli (2007), 100000000 is returned when any \eqn{\sigma_i = 0}
due to its application in the optimisation of \eqn{\beta} for diffusion kernel k-means clustering.
It avoids \eqn{\beta} that gives eigenvalues of 0 in the computation of the diffusion kernel.
}
\references{
Kondor, R. I., & Lafferty, J. (2002, July). Diffusion kernels on graphs and other discrete structures. In Proceedings of the 19th international conference on machine learning (Vol. 2002, pp. 315-322).

Kondor, R., & Vert, J. P. (2004). Diffusion kernels. kernel methods in computational biology, 171-192.

Roy, O., & Vetterli, M. (2007, September). The effective rank: A measure of effective dimensionality. In 2007 15th European signal processing conference (pp. 606-610). IEEE.
}
