% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{compute_laplacian}
\alias{compute_laplacian}
\title{Compute graph Laplacian}
\usage{
compute_laplacian(W, normalise = TRUE)
}
\arguments{
\item{W}{a symmetric adjaceny matrix.}

\item{normalise}{a Boolean to normalise the graph Laplacian. Default is TRUE.}
}
\value{
a graph Laplacian matrix.
}
\description{
This function computes and returns a graph Laplacian.
}
\references{
Chung, F. R. (1997). Spectral graph theory (Vol. 92). American Mathematical Soc..

Smola, A. J., & Kondor, R. (2003, August). Kernels and regularization on graphs. In Learning theory and kernel machines: 16th annual conference on learning theory and 7th kernel workshop, COLT/kernel 2003, Washington, DC, USA, august 24-27, 2003. Proceedings (pp. 144-158). Berlin, Heidelberg: Springer Berlin Heidelberg.
}
