% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{compute_spacecost}
\alias{compute_spacecost}
\title{Spatial cost computation}
\usage{
compute_spacecost(space_kmat, clust_points, i = NULL)
}
\arguments{
\item{space_kmat}{a kernel matrix computed from the spatial distance matrix.}

\item{clust_points}{a numeric vector of point indices in the targeted cluster.}

\item{i}{an integer of a specific point index. Default is NULL.}
}
\value{
a numeric value or vector of spatial cost(s) for a given cluster.
}
\description{
This function computes and returns the spatial costs of all points for a given cluster.
}
\references{
Dhillon, I. S., Guan, Y., & Kulis, B. (2004, August 22). Kernel k-means: spectral clustering and normalized cuts. Proceedings of the Tenth ACM SIGKDD International Conference on Knowledge Discovery and Data Mining. KDD04: ACM SIGKDD International Conference on Knowledge Discovery and Data Mining, Seattle WA USA. https://doi.org/10.1145/1014052.1014118

Schölkopf, B., Smola, A., & Müller, K.-R. (1998). Nonlinear component analysis as a kernel eigenvalue problem. Neural Computation, 10(5), 1299–1319.
}
