% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{compute_spacecost_kmedoids}
\alias{compute_spacecost_kmedoids}
\title{Compute k-medoids spatial costs}
\usage{
compute_spacecost_kmedoids(
  space_distmat,
  clust_points,
  i = NULL,
  weights = NULL
)
}
\arguments{
\item{space_distmat}{a spatial distance matrix.}

\item{clust_points}{a numeric vector of point indices in the targeted cluster.}

\item{i}{an integer of a specific point index. Default is NULL.}

\item{weights}{a numeric vector of weights for each data point. Default is NULL.}
}
\value{
a numeric value or vector of the spatial cost(s) for a given cluster.
}
\description{
This function computes and returns the spatial costs of all points for a given cluster using k-medoids clustering.
}
