% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{compute_spacestat}
\alias{compute_spacestat}
\title{Spatial cost computation}
\usage{
compute_spacestat(space_distmat, clust_points, sigma)
}
\arguments{
\item{space_distmat}{a numeric spatial distance matrix.}

\item{clust_points}{a numeric vector of point indices in the targeted cluster.}

\item{sigma}{a numeric value of sigma passed on to \code{\link[=rbf]{rbf()}}.}
}
\value{
a numeric vector of spatial statistics for a given cluster.
}
\description{
This function computes and returns the statistic of the spatial cost function for a given cluster.
}
