% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{diffusion_kernel}
\alias{diffusion_kernel}
\title{Compute diffusion kernel}
\usage{
diffusion_kernel(L, beta)
}
\arguments{
\item{L}{a graph Laplacian.}

\item{beta}{a numeric parameter controlling the diffusion rate.}
}
\value{
a diffusion kernel matrix
}
\description{
This function computes and returns a kernel matrix using a diffusion kernel as defined by Kondor and Lafferty (2002).
}
\references{
Kondor, R. I., & Lafferty, J. (2002, July). Diffusion kernels on graphs and other discrete structures. In Proceedings of the 19th international conference on machine learning (Vol. 2002, pp. 315-322).

Kondor, R., & Vert, J. P. (2004). Diffusion kernels. kernel methods in computational biology, 171-192.
}
