% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{distmat_to_kmat}
\alias{distmat_to_kmat}
\title{Convert distance matrix to kernel matrix using diffusion kernel}
\usage{
distmat_to_kmat(D, beta)
}
\arguments{
\item{D}{a distance matrix.}

\item{beta}{a numeric parameter controlling the diffusion rate.}
}
\value{
a diffusion kernel matrix
}
\description{
This function converts a distance matrix to a kernel matrix using a diffusion kernel.
}
\details{
A weighted adjacency matrix \eqn{W} is obtained from the distance matrix using the radial basis function (i.e. Gaussian function).
Here, we fix the hyperparameter \eqn{\sigma} as the median of the distance distribution.
Then, an unnormalised graph Laplacian \eqn{L} is constructed by \eqn{L = D - W} where \eqn{D} is the degree matrix obtained from \eqn{W}.
A diffusion kernel is used to compute the kernel matrix by \eqn{K = e^{-\beta L}} via eigendecomposition.
}
