% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{eval_blobs}
\alias{eval_blobs}
\title{Clustering evaluation}
\usage{
eval_blobs(data, crs = 4326, ...)
}
\arguments{
\item{data}{a data frame.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}. Default is 4326.}

\item{...}{the additional arguments include \code{k_matrix}, \code{beta} and \code{space_distmat}.
\itemize{
\item \code{k_matrix}: a kernel matrix computed from the spatial distance matrix.
\item \code{beta}: a numeric parameter controlling the diffusion rate passed on to \code{\link[=k_diffusion]{k_diffusion()}}.
\item \code{space_distmat}: a numeric spatial distance matrix.
}}
}
\value{
a list of the following objects.
\itemize{
\item \code{summary}: a data frame of summary statistics.
\item \code{clust_below_size}: a numeric vector of clusters below the critical size.
}
}
\description{
This function evaluates clustering performance and returns a summary and clusters that are below the critical size.
}
\details{
The critical size of a cluster is defined as \eqn{\frac{N}{2K}} where \eqn{N} is the number of data points and \eqn{k} is the number of clusters.
}
