% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moo.R
\name{eval_moo}
\alias{eval_moo}
\title{Evaluate multi-objective optimisation performance}
\usage{
eval_moo(batch_list, obj)
}
\arguments{
\item{batch_list}{a list of \code{pop} objects from different batches; see also \code{\link[=blob_populate_batch]{blob_populate_batch()}}.}

\item{obj}{a string vector of objectives. "-" prefix to indicate objectives to be maximised.}
}
\value{
a \code{pop_pareto} object includes a list of the following objects.
\itemize{
\item \code{clust}: a numeric matrix of cluster assignments. Each row is a Pareto optimal solution.
\item \code{summary}: a data frame of summary statistics of all feasible solutions.
\item \code{trace}: a data frame of summary statistics for tracing of all feasible solutions.
\item \code{n_filtered}: a data frame of numbers of filtered solutions.
\item \code{space_kmat_optim_out}: an output of \code{\link[stats:optim]{stats::optim()}} from the optimisation of \eqn{\beta} in \code{\link[=distmat_to_kmat]{distmat_to_kmat()}} when \code{space_kmat} is not supplied.
\item \code{pareto_idx}: a numeric vector of indices of Pareto optimal cluster assignment.
\item \code{obj}: a string vector of objectives.
\item \code{moo_quality}: a data frame of MOO quality indicators.
}
}
\description{
This function evaluates multi-objective optimisation (MOO) performance.
}
\details{
The quality indicators include IGD, IGD+ and hypervolume commonly used in MOO,
computed using \code{\link[moocore:igd]{moocore::igd()}}, \code{\link[moocore:igd]{moocore::igd_plus()}}, and \code{\link[moocore:hypervolume]{moocore::hypervolume()}}.

The reference for IGD and IGD+ are the Pareto optimal solutions from the final cumulative batch and
the reference point in hypervolume is \eqn{(1_m = (1,1,,1) \in \R^m)} for \eqn{m} objectives.

The lower and upper bounds for normalisation are obtained from all the solutions.

See \link[moocore:moocore-package]{moocore::moocore} for more information.
}
\seealso{
\code{\link[=combine_pop]{combine_pop()}}, \code{\link[=find_pareto]{find_pareto()}}, \link[moocore:moocore-package]{moocore::moocore}
}
