% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{filter_pareto_similar}
\alias{filter_pareto_similar}
\title{Filter similar Pareto optimal solutions}
\usage{
filter_pareto_similar(pop, ari)
}
\arguments{
\item{pop}{a list of objects returned from \code{\link[=blob_populate]{blob_populate()}} or \code{\link[=blob_kpopulate]{blob_kpopulate()}} from the current batch.}

\item{ari}{a numeric value of Adjusted Rand Index (ARI) that sets duplication threshold between two assignments. It must be \eqn{[0,1]}. Default is 1. See also \code{\link[mclust:adjustedRandIndex]{mclust::adjustedRandIndex()}}.}
}
\value{
a list of the following objects.
\itemize{
\item \code{clust}: a numeric matrix of cluster assignments. Each row is a Pareto optimal solution.
\item \code{summary}: a data frame of summary statistics.
\item \code{trace}: a data frame of summary statistics for tracing.
\item \code{n_filtered}: a data frame of numbers of filtered solutions.
\item \code{clust_idx}: a numeric vector of indices of cluster assignment.
\item \code{obj}: a string vector of objectives.
}
}
\description{
This function filters Pareto optimal solutions and returns an updated \code{\link[=blob_moo]{blob_moo()}} output.
The order of \code{obj} indicates the importance in ranking the solutions when considering which solutions to keep in the face of similarity.
}
