% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{find_blobs}
\alias{find_blobs}
\title{Assign clusters}
\usage{
find_blobs(data, k, r, ...)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters.}

\item{r}{an numeric value of spatial relative weight. It must be \eqn{[0,1]}.}

\item{...}{the additional arguments include \code{k_matrix}, \code{beta} and \code{space_distmat}.
\itemize{
\item \code{k_matrix}: a kernel matrix computed from the spatial distance matrix.
\item \code{beta}: a numeric parameter controlling the diffusion rate passed on to \code{\link[=k_diffusion]{k_diffusion()}}.
\item \code{space_distmat}: a numeric spatial distance matrix.
}}
}
\value{
a data frame with assigned clusters as a column.
}
\description{
This function assigns clusters for a given k and r.
}
\details{
Diffusion kernel is applied to compute distance to centroid in space.
}
