% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{find_blobs_distmat}
\alias{find_blobs_distmat}
\title{Assign clusters - spatial distances supplied by a distance matrix}
\usage{
find_blobs_distmat(data, k, r, space_distmat, sigma)
}
\arguments{
\item{data}{a data matrix or data frame with assigned (starting) clusters as a column.}

\item{k}{an integer of the number of clusters.}

\item{r}{an numeric value of spatial relative weight. It must be \eqn{[0,1]}.}

\item{space_distmat}{a numeric spatial distance matrix. Default is NULL.}
}
\value{
a data frame with assigned clusters as a column.
}
\description{
This function assigns clusters for a given k and r. Spatial distances are supplied by a distance matrix.
}
\details{
When distances are supplied by a distance matrix, kernel method is applied to compute cluster centroids in space.
}
