% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{find_blobs_otf}
\alias{find_blobs_otf}
\title{Assign clusters - spatial distances computed on the fly}
\usage{
find_blobs_otf(data, k, r, space_distmethod = "geodesic")
}
\arguments{
\item{data}{a data matrix or data frame with assigned (starting) clusters as a column.}

\item{k}{an integer of the number of clusters.}

\item{r}{an numeric value of spatial relative weight. It must be \eqn{[0,1]}.}

\item{space_distmethod}{a string of the method for distance computation. It must be either "geodesic" or "euclidean". Default is "geodesic".}
}
\value{
a data frame with assigned clusters as a column.
}
\description{
This function assigns clusters for a given k and r. Spatial distances are computed on the fly.
}
\details{
When distances are computed on the fly, kernel method cannot be applied to compute cluster centroids in space.
}
