% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{find_dup}
\alias{find_dup}
\title{Find duplicates}
\usage{
find_dup(clust, ari = 1)
}
\arguments{
\item{clust}{a numeric matrix of cluster assignments. Each row is an assignment.}

\item{ari}{a numeric value of Adjusted Rand Index (ARI) that sets duplication threshold between two assignments. It must be \eqn{[0,1]}. Default is 1. See also \code{\link[mclust:adjustedRandIndex]{mclust::adjustedRandIndex()}}.}
}
\value{
a list of the following objects.
\itemize{
\item \code{idx}: a numeric vector of duplicate indices.
\item \code{pairwise_ari}: a data frame of summary statistics.
\item \code{freq}: a contingency table of the frequency of duplicates, an object of class "table", an array of integer values. See also \code{\link[base:table]{base::table()}}.
\item \code{pairs_dup}: a numeric matrix of pairs of duplicates. The second row stores indices of duplicates of the first row.
}
}
\description{
This function finds duplicates in a set of cluster assignments. It uses \code{\link[mclust:adjustedRandIndex]{mclust::adjustedRandIndex()}} to measure similarity.
}
\seealso{
\code{\link[base:table]{base::table()}}, \code{\link[mclust:adjustedRandIndex]{mclust::adjustedRandIndex()}}
}
