% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_moo.R
\name{find_pareto}
\alias{find_pareto}
\title{Find Pareto optimal solutions}
\usage{
find_pareto(pop, obj = NULL)
}
\arguments{
\item{pop}{a \code{pop} object; see also \code{\link[=blob_populate]{blob_populate()}}.}

\item{obj}{a string vector of objectives. "-" prefix to indicate objectives to be maximised.}
}
\value{
a \code{pop_pareto} object includes a list of the following objects.
\itemize{
\item \code{clust}: a numeric matrix of the cluster assignments. Each row is a Pareto optimal solution.
\item \code{summary}: a data frame of the summary statistics.
\item \code{trace}: a data frame of the summary statistics per iteration.
\item \code{n_filtered}: a data frame of the numbers of filtered solutions.
\item \code{space_kmat_optim_out}: an output of \code{\link[stats:optim]{stats::optim()}} from the optimisation of \eqn{\beta} in \code{\link[=distmat_to_kmat]{distmat_to_kmat()}} when \code{space_kmat} is not supplied.
\item \code{pareto_idx}: a numeric vector of indices of Pareto optimal cluster assignment.
\item \code{obj}: a string vector of objectives.
}
}
\description{
This function finds a set of Pareto optimal solutions from a given set of solutions using \code{\link[moocore:nondominated]{moocore::is_nondominated()}} and updates \code{pop}.
}
\seealso{
\code{\link[moocore:nondominated]{moocore::is_nondominated()}}
}
