% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{gen_gaussian_data}
\alias{gen_gaussian_data}
\title{Generate Gaussian data clusters}
\usage{
gen_gaussian_data(n, size, center = NULL, ...)
}
\arguments{
\item{n}{an integer of the number of clusters}

\item{size}{an integer vector of sizes of a cluster. If an integer value is supplied, all clusters will be of this size.}

\item{center}{a data matrix or data frame of coordinates.}

\item{...}{the optional argument includes \code{sigma}.
\itemize{
\item \code{sigma}: a covariance matrix or a list of covariance matrices to specify different covariance matrices. See \code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm()}}.
}}
}
\value{
a data frame of coordinates with clusters as a column.
}
\description{
This function generates Gaussian data clusters.
}
\seealso{
\code{\link[mvtnorm:Mvnorm]{mvtnorm::rmvnorm()}}
}
