% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{intersects_bool}
\alias{intersects_bool}
\title{Intersect check}
\usage{
intersects_bool(data, clust = NULL, crs = 4326)
}
\arguments{
\item{data}{a numeric data matrix or data frame.}

\item{clust}{a numeric vector of cluster assignment. Default is NULL when it is already present as a column in the data.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}. Default is 4326.}
}
\value{
TRUE or FALSE
}
\description{
This function checks and indicates if any clusters are intersecting based on their convex hulls.
}
