% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jain}
\alias{jain}
\title{"Jain" data}
\format{
A data frame with 373 rows and 4 columns:
\describe{
\item{x, y}{Spatial coordinates of the original dataset}
\item{age}{Randomly sampled age}
\item{clust}{True spatial cluster}
}
}
\source{
\url{https://cs.joensuu.fi/sipu/datasets/}
}
\usage{
jain
}
\description{
"Jain" data
}
\references{
Jain, A. K., & Law, M. H. (2005, December). Data clustering: A user’s dilemma. In International conference on pattern recognition and machine intelligence (pp. 1-10). Berlin, Heidelberg: Springer Berlin Heidelberg.

Fränti, P., & Sieranoja, S. (2018). K-means properties on six clustering benchmark datasets. Applied intelligence, 48(12), 4743-4759.
}
\keyword{datasets}
