% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{k_diffusion}
\alias{k_diffusion}
\title{Compute diffusion kernel}
\usage{
k_diffusion(L, beta)
}
\arguments{
\item{L}{a graph Laplacian}

\item{beta}{a numeric parameter controlling the diffusion rate.}
}
\value{
a diffusion kernel matrix
}
\description{
This function compute a diffusion kernel using eigendecomposition.
}
