% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{min_erankdiff}
\alias{min_erankdiff}
\title{Loss function to minimise the difference in effective ranks}
\usage{
min_erankdiff(beta, L, k)
}
\arguments{
\item{beta}{a numeric parameter controlling the diffusion rate.}

\item{L}{a graph Laplacian.}

\item{k}{an integer of the targeted effective rank.}
}
\value{
a numeric value of the difference in effective ranks.
}
\description{
This function calculates and returns the absolute difference between the targeted effective rank and
the effective rank of the diffusion kernel for a given \code{beta} (\eqn{\beta}) based on the effective rank defined by Roy and Vetterli (2007).
}
\details{
As the diffusion kernel is symmetric and positive definite,
its eigenvalues, defined as  \eqn{e^{\beta\lambda_i}},
are essentially singular values \eqn{\sigma_i}, where \eqn{\lamba_i} are eigenvalues of\eqn{L}.
This is utilised to calculate the effective rank of the kernel for a given \eqn{\beta} following Roy and Vetterli (2007) and
this loss function to be minimised is defined as the absolute difference between that and the targeted effective rank.

Instead of treating \eqn{0\log{0} = 0}
when any singular values \eqn{\sigma_i = 0} in Roy and Vetterli (2007),
1e8 is returned for its application in the optimisation of the diffusion rate (\eqn{\beta}) for kernel k-means clustering
with the constraint \eqn{e^{\beta\lambda_i}} > 0.
}
\references{
Kondor, R. I., & Lafferty, J. (2002, July). Diffusion kernels on graphs and other discrete structures. In Proceedings of the 19th international conference on machine learning (Vol. 2002, pp. 315-322).

Kondor, R., & Vert, J. P. (2004). Diffusion kernels. kernel methods in computational biology, 171-192.

Roy, O., & Vetterli, M. (2007, September). The effective rank: A measure of effective dimensionality. In 2007 15th European signal processing conference (pp. 606-610). IEEE.
}
