% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noisycircles}
\alias{noisycircles}
\title{"Noisycircles" dataset}
\format{
\subsection{\code{noisycircles}}{

A data frame with 1000 rows and 4 columns:
\describe{
\item{x, y}{Spatial coordinates of the original dataset}
\item{age}{Randomly sampled age}
\item{clust}{True spatial cluster}
}
}
}
\source{
\link[factoextra:multishapes]{factoextra::multishapes}
}
\usage{
noisycircles
}
\description{
"Noisycircles" dataset
}
\keyword{datasets}
