% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{optim_beta}
\alias{optim_beta}
\title{Optimise beta for diffusion kernel}
\usage{
optim_beta(L, k, par = 10)
}
\arguments{
\item{L}{a graph Laplacian.}

\item{k}{an integer of the number of clusters.}

\item{par}{an numeric initial value for the parameter to be optimised over. The range is \eqn{[0.001, \inf]}. Default is 10.}
}
\value{
For \code{optim}, a list with components:
  \item{par}{The best set of parameters found.}
  \item{value}{The value of \code{fn} corresponding to \code{par}.}
  \item{counts}{A two-element integer vector giving the number of calls
    to \code{fn} and \code{gr} respectively. This excludes those calls needed
    to compute the Hessian, if requested, and any calls to \code{fn} to
    compute a finite-difference approximation to the gradient.}
  \item{convergence}{An integer code. \code{0} indicates successful
    completion (which is always the case for \code{"SANN"} and
    \code{"Brent"}).  Possible error codes are
    \describe{
      \item{\code{1}}{indicates that the iteration limit \code{maxit}
      had been reached.}
      \item{\code{10}}{indicates degeneracy of the Nelder--Mead simplex.}
      \item{\code{51}}{indicates a warning from the \code{"L-BFGS-B"}
      method; see component \code{message} for further details.}
      \item{\code{52}}{indicates an error from the \code{"L-BFGS-B"}
      method; see component \code{message} for further details.}
    }
  }
  \item{message}{A character string giving any additional information
    returned by the optimizer, or \code{NULL}.}
  \item{hessian}{Only if argument \code{hessian} is true. A symmetric
    matrix giving an estimate of the Hessian at the solution found.  Note
    that this is the Hessian of the unconstrained problem even if the
    box constraints are active.}

  For \code{optimHess}, the description of the \code{hessian} component
  applies.
}
\description{
This function optimises \eqn{\beta} for the diffusion kernel by minimising the absolute difference between
the number of clusters and the effective rank of the kernel matrix.
}
\details{
To perform diffusion kernel k-means clustering, the diffusion rate (\eqn{\beta}) has to be optimised to capture the global similarity of
a graph Lapclacian matrix \eqn{L} in the feature space.

This is done by minimising the absolute difference between the number of clusters and the effective rank of the kernel matrix where the effective rank,
which is a measure of effective dimensionality of a matrix, follows the definition in Roy and Vetterli (2007).

As the diffusion kernel is symmetric and positive definite,
its eigenvalues \eqn{e^{\beta\lambda_i}}
are essentially singular values \eqn{\sigma_i}, where \eqn{\lamdba_i} are eigenvalues of\eqn{L}.
This is utilised to calculate the effective rank of the kernel matrix for a given \eqn{\beta}
and the optimisation is performed by \code{\link[stats:optim]{stats::optim()}} using "L-BFGS-B" method with a lower bound of 0.001.
}
\references{
Roy, O., & Vetterli, M. (2007, September). The effective rank: A measure of effective dimensionality. In 2007 15th European signal processing conference (pp. 606-610). IEEE.

Kondor, R. I., & Lafferty, J. (2002, July). Diffusion kernels on graphs and other discrete structures. In Proceedings of the 19th international conference on machine learning (Vol. 2002, pp. 315-322).

Kondor, R., & Vert, J. P. (2004). Diffusion kernels. kernel methods in computational biology, 171-192.
}
\seealso{
\code{\link[=compute_erank]{compute_erank()}}, \code{\link[stats:optim]{stats::optim()}}
}
