% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{parse_obj}
\alias{parse_obj}
\title{Parse objective parameter}
\usage{
parse_obj(obj)
}
\arguments{
\item{obj}{a string vector of objectives.}
}
\value{
a list of the following objects.
\itemize{
\item \code{obj}: a string vector of parsed objectives.
\item \code{maximise_obj_idx}: an numeric vector of indices of the objective to be maximised.
}
}
\description{
This function parses \code{obj} string vector for \code{\link[=blob_moo]{blob_moo()}}.
}
\details{
The leading "-" (minus) is used to indicate maximisation of the objective.
}
