% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{pivot_trace}
\alias{pivot_trace}
\title{Pivot trace summary for plotting}
\usage{
pivot_trace(df)
}
\arguments{
\item{df}{a data frame of trace summary.}
}
\value{
a pivoted data frame of trace summary.
}
\description{
This function pivots trace summary for plotting.
}
\seealso{
\code{\link[tidyr:pivot_longer]{tidyr::pivot_longer()}}
}
