% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_2d_objspace}
\alias{plot_2d_objspace}
\title{Plot 2D objective space}
\usage{
plot_2d_objspace(
  pop,
  obj,
  colour = c("r", "batch", "k_o", "pareto"),
  normalise = T,
  palette = NULL,
  alpha = 0.8
)
}
\arguments{
\item{pop}{a list of objects returned from \code{\link[=blob_moo]{blob_moo()}}.}

\item{obj}{a string vector of length 2 that indicate the objectives.}

\item{colour}{a string of the attribute to be coloured. It must be "pareto", "r", "batch" or "k_o".}

\item{normalise}{a logical operator. Should values of objectives be normalised? Default is T.}

\item{palette}{See \code{value} of \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}. Default is NULL. This uses the palette presets from \code{\link[MetBrewer:met.brewer]{MetBrewer::met.brewer()}}.}

\item{alpha}{See <\code{\link[ggplot2:aes_colour_fill_alpha]{aes-colour-fill-alpha}}>.}
}
\value{
a ggplot.
}
\description{
This function plots 2D objective space.
}
\seealso{
\code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}, \code{\link[MetBrewer:met.brewer]{MetBrewer::met.brewer()}}
}
