% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_moo_quality}
\alias{plot_moo_quality}
\title{Plot multi-objective optimisation quality}
\usage{
plot_moo_quality(pop, indicator)
}
\arguments{
\item{pop}{a list of objects returned from \code{\link[=blob_moo]{blob_moo()}}.}

\item{indicator}{a string of MOO quality indicator. It must be either "igd", "igd_plus" or "hv".}
}
\value{
a ggplot.
}
\description{
This function plots multi-objective optimisation (MOO) quality.
}
