% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_mooquality}
\alias{plot_mooquality}
\title{Plot multi-objective optimisation quality}
\usage{
plot_mooquality(pop_moo, indicator = c("igd", "igd_plus", "hv"))
}
\arguments{
\item{pop_moo}{a \code{pop_moo} object; see also \code{\link[=blob_moo]{blob_moo()}}.}

\item{indicator}{a string of MOO quality indicator. It must be either "igd", "igd_plus" or "hv".}
}
\value{
a \code{ggplot} object.
}
\description{
This function plots multi-objective optimisation (MOO) quality.
}
