% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_objspace}
\alias{plot_objspace}
\title{Plot objective space}
\usage{
plot_objspace(
  pop,
  obj,
  colour = c("r", "batch", "k_o", "pareto"),
  normalise = TRUE,
  alpha = 0.8
)
}
\arguments{
\item{pop}{a list of objects returned from \code{\link[=blob_moo]{blob_moo()}}.}

\item{obj}{a string vector of length 2 that indicates the objectives.}

\item{colour}{a string of the attribute to be coloured. It must be either "pareto", "r", "batch" or "k_o".}

\item{normalise}{a Boolen. Should values of objectives be normalised? Default is TRUE.}

\item{alpha}{See <\code{\link[ggplot2:aes_colour_fill_alpha]{aes-colour-fill-alpha}}>.}
}
\value{
a \code{ggplot} object.
}
\description{
This function plots the objective space.
}
