% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_pcp}
\alias{plot_pcp}
\title{Plot parallel coordinate plot}
\usage{
plot_pcp(
  pop,
  colour = c("r", "k", "batch", "pareto"),
  normalise = TRUE,
  pareto = TRUE,
  pareto_similar = FALSE,
  linewidth = 1,
  alpha = 0.8
)
}
\arguments{
\item{pop}{a list of objects returned from \code{\link[=stblob]{stblob()}}.}

\item{colour}{a string of the attribute to be coloured. It must be either "pareto", "r", "batch" or "k".}

\item{normalise}{a Boolen. Should values of objectives be normalised? Default is TRUE.}

\item{pareto}{a Boolen. Should only the Pareto solutions be displayed? Default is TRUE.}

\item{pareto_similar}{a Boolen. Should only the similar Pareto solutions be displayed? Default is FALSE.}

\item{linewidth}{linewidth for \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.}

\item{alpha}{See <\code{\link[ggplot2:aes_colour_fill_alpha]{aes-colour-fill-alpha}}>.}
}
\value{
a \code{ggplot} object.
}
\description{
This function plots the objective space.
}
