% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_time}
\alias{plot_time}
\title{Plot data in time}
\usage{
plot_time(data, clust, age = 3)
}
\arguments{
\item{data}{a data frame}

\item{clust}{a numeric vector of cluster assignment.}

\item{age}{a string or numeric value indicating the column of age. Default is the third column.}
}
\description{
This function plots data in time
}
