% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot trace}
\usage{
plot_trace(pop, colour = c("r", "batch", "k"), alpha = 0.8)
}
\arguments{
\item{pop}{a list of objects returned from \code{\link[=stblob]{stblob()}}.}

\item{colour}{a string of the attribute to be coloured. It must be either "r", "batch" or "k".}

\item{alpha}{See <\code{\link[ggplot2:aes_colour_fill_alpha]{aes-colour-fill-alpha}}>.}
}
\value{
a ggplot.
}
\description{
This function plots the trace.
}
\seealso{
\code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}, \code{\link[MetBrewer:met.brewer]{MetBrewer::met.brewer()}}
}
