% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{pop_update}
\alias{pop_update}
\title{Update "pop"}
\usage{
pop_update(pop_old, pop)
}
\arguments{
\item{pop_old}{a list of objects returned from \code{\link[=blob_populate]{blob_populate()}} or \code{\link[=blob_kpopulate]{blob_kpopulate()}} from the previous batch.}

\item{pop}{a list of objects returned from \code{\link[=blob_populate]{blob_populate()}} or \code{\link[=blob_kpopulate]{blob_kpopulate()}} from the current batch.}
}
\value{
a list of the following objects.
\itemize{
\item \code{clust}: a numeric matrix of cluster assignments. Each row is a solution.
\item \code{summary}: a data frame of summary statistics.
\item \code{trace}: a data frame of summary statistics for tracing.
\item \code{n_filtered}: a data frame of numbers of filtered solutions.
}
}
\description{
This function update and append the current batch to a list of returned objects from \code{\link[=blob_populate]{blob_populate()}} or \code{\link[=blob_kpopulate]{blob_kpopulate()}}, from the previous batch.
}
