% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{rbf}
\alias{rbf}
\title{Radial basis function}
\usage{
rbf(D, sigma)
}
\arguments{
\item{D}{a distance matrix.}

\item{sigma}{a numeric value of sigma.}
}
\value{
a weighted adjacency matrix \eqn{[0,1]}.
}
\description{
This function converts a distance matrix to a weighted adjacency matrix using the radial basis function (i.e. Gaussian function).
}
