% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{savePDF}
\alias{savePDF}
\title{Save a plot as PDF}
\usage{
savePDF(p, file, width, height, ...)
}
\arguments{
\item{p}{plot.}

\item{...}{
  Arguments passed on to \code{\link[grDevices:pdf]{grDevices::pdf}}
  \describe{
    \item{\code{file}}{a character string giving the file path.
    See the section \sQuote{File specifications} for further details.
  }
    \item{\code{width,height}}{the width and height of the graphics region in
    inches.  The default values are \code{7}.}
  }}
}
\description{
This function saves a plot as PDF. As \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} can be buggy with text spacing.
}
\seealso{
\code{\link[grDevices:pdf]{grDevices::pdf()}}
}
