% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{slc}
\alias{slc}
\title{Geodesic distance computation}
\usage{
slc(x, y, ax, ay, input = "rad")
}
\arguments{
\item{x}{a numeric value of longitude.}

\item{y}{a numeric value of latitude.}

\item{ax}{a numeric vector of longitudes.}

\item{ay}{a numeric vector of latitudes.}
}
\value{
a numeric vector of distances in radians.
}
\description{
This function calculates and returns the shortest distance in radians between a single (x,y) and all points (xa,ya) using spherical law of cosines.
}
