% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{start_blobs_distmat}
\alias{start_blobs_distmat}
\title{Assign the starting cluster members - spatial distances supplied by a distance matrix}
\usage{
start_blobs_distmat(data, k, space_distmat)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters.}

\item{space_distmat}{a numeric spatial distance matrix.}
}
\value{
a data frame with assigned starting clusters as a column.
}
\description{
Assign the starting cluster members by approximating the maximum spatial distant points. The distances are supplied by a distance matrix.
}
