% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{start_blobs_otf}
\alias{start_blobs_otf}
\title{Assign the starting cluster members - spatial distances computed on the fly}
\usage{
start_blobs_otf(data, k, space_distmethod = "geodesic")
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters}

\item{space_distmethod}{a string of the method for distance computation. It must be either "geodesic" or "euclidean". Default is "geodesic".}
}
\value{
a data frame with assigned starting clusters as a column.
}
\description{
Assign the starting cluster members by approximating the maximum spatial distant points. The distances are computed on the fly.
}
