% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{threeblobs}
\alias{threeblobs}
\title{Demo data}
\format{
A data frame with 300 rows and 4 columns:
\describe{
\item{x, y}{Spatial coordinates of the original dataset}
\item{age}{Ordered age for demomstration}
\item{clust}{Simulated spatial cluster}
}
}
\source{
This package
}
\usage{
threeblobs
}
\description{
Demo data
}
\keyword{datasets}
