% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{threeblobs_a}
\alias{threeblobs_a}
\title{Demo data A}
\format{
A data frame with 300 rows and 4 columns:
\describe{
\item{x, y}{Spatial coordinates of the original dataset}
\item{age}{Randomly sampled age}
\item{clust}{True spatial cluster}
}
}
\source{
This package
}
\usage{
threeblobs_a
}
\description{
Demo data A
}
\keyword{datasets}
