% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{blob_search}
\alias{blob_search}
\title{Core local-search algorithm}
\usage{
blob_search(
  data,
  k,
  r,
  iter = 10,
  converge_ari = NULL,
  coords = c(1, 2),
  age = 3,
  crs = 4326,
  hull_convex_ratio = 0,
  random_start = FALSE,
  filter_intersects = TRUE,
  filter_clustsize = TRUE,
  max_na = 0.05,
  space_kmat = NULL,
  space_distmat = NULL,
  space_distmethod = NULL,
  w_knn = NULL,
  l_normalise = NULL,
  beta_par = NULL
)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters.}

\item{r}{an numeric value of spatial relative weight. It must be \eqn{[0,1]}.}

\item{iter}{an integer of the number of iterations. Default is 10.}

\item{converge_ari}{a numeric value of Adjusted Rand Index (ARI) that sets convergence threshold between two searches. It must be \eqn{[0,1]}. Default is NULL.}

\item{coords}{a vector of strings or numeric values indicating the columns of coordinates (longitude, latitide). Default is the first two columns.}

\item{age}{a string or numeric value indicating the column of age. Default is the third column.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}. Default is 4326.}

\item{hull_convex_ratio}{a numeric value controlling the convexity of the hulls passed onto \code{\link[sf:geos_unary]{sf::st_concave_hull()}}. 1 returns convex and 0 maximally concave hulls. Default is 0.}

\item{random_start}{a Boolean to use random starting cluster members. Default is FALSE.}

\item{filter_intersects}{a Boolean to remove an assignment with intersects in space? Default is TRUE.}

\item{filter_clustsize}{a Boolean to assign NA to clustes below the critical size. Default is TRUE.}

\item{max_na}{a numeric value of the maximum proportion of NAs allowed. It must be \eqn{[0,1]}. Default is 0.05.}

\item{space_kmat}{a kernel matrix computed from the spatial distance matrix.}

\item{space_distmat}{a spatial distance matrix used when \code{space_kmat} is not supplied. Default is NULL.}

\item{space_distmethod}{a string of method used when \code{space_kmat} and \code{space_distmat} are not specified. It must be one of "geodesic" or "euclidean". Default is NULL.}

\item{w_knn}{an integer of the \eqn{k^{th}} nearest neighbour used when \code{space_kmat} used in local scaling for computing the adjacency matrix, passed onto \code{\link[=compute_adjacency]{compute_adjacency()}}. Default is 7 when \code{space_kmat} is not supplied.}

\item{l_normalise}{a Boolean to normalise graph Laplacian, passed onto \code{\link[=compute_laplacian]{compute_laplacian()}}. Default is TRUE when \code{space_kmat} is not supplied.}

\item{beta_par}{an numeric initial value for the parameter \eqn{\beta} to be optimised over for the diffusion kernel, passed onto \code{\link[=optim_beta]{optim_beta()}}. The range is \eqn{[0.001, \inf]}. Default is 10 when \code{space_kmat} is not supplied.}
}
\value{
a numeric value of 1 if the search result is removed due to the presence of intersects,
2 if removed due to the proportion of NAs exceeding \code{max_na},
3 if removed due to 1 resulting cluster,
or a list of the following objects if the search is meets the constraints,
\itemize{
\item \code{data}: a data frame of the input data with assigned clusters as a column.
\item \code{summary}: a data frame of summary statistics.
\item \code{clust_below_size}: a numeric vector of clusters below the critical size.
\item \code{trace}: a data frame of summary statistics for tracing.
\item \code{space_kmat_optim_out}: an output of \code{\link[stats:optim]{stats::optim()}} from the optimisation of \eqn{\beta} in \code{\link[=distmat_to_kmat]{distmat_to_kmat()}} when \code{space_kmat} is not supplied.
}
}
\description{
This function performs an iterative bi-objective local search algorithm to assign clusters for a given k and r.
}
\details{
A diffusion kernel is applied to compute the distance to the centroid in space.
See \code{\link[=distmat_to_kmat]{distmat_to_kmat()}} for more details for converting a distance matrix to a kernel matrix.

Clusters are assigned in every iteration. It iterates until the set length or convergence.

When \code{converge_ari} is specified, convergence is defined and activated when ARI between the latest and the previous search is
above the specified threshold and at least three iterations are run.

The critical size of a cluster is defined as \eqn{\frac{n}{2k}} where \eqn{n} is the number of data point and \eqn{k} is the number of clusters.
}
\seealso{
\code{\link[=distmat_to_kmat]{distmat_to_kmat()}}, \code{\link[sf:st_as_sf]{sf::st_as_sf()}}
}
