% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{compute_distmat}
\alias{compute_distmat}
\title{Compute distance matrix}
\usage{
compute_distmat(data, method = "geodesic")
}
\arguments{
\item{data}{a numeric matrix or data frame.}

\item{method}{a string of the method. Either "geodesic" or "euclidean" is available. Default is "geodesic".}
}
\value{
a numeric distance matrix.
}
\description{
This function computes and returns a distance matrix.
}
\details{
km is the unit when "geodesic" is specified.
}
