% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{compute_laplacian}
\alias{compute_laplacian}
\title{Compute unnormalised graph Laplacian}
\usage{
compute_laplacian(A)
}
\arguments{
\item{A}{an adjaceny matrix}
}
\value{
an unnormalised graph Laplacian.
}
\description{
This function compute unnormalised graph Laplacian.
}
