% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{compute_spacestat}
\alias{compute_spacestat}
\title{Spatial cost computation}
\usage{
compute_spacestat(clust_points, ...)
}
\arguments{
\item{clust_points}{a numeric vector of point indices in the targeted cluster.}

\item{...}{the additional arguments include \code{k_matrix}, \code{beta} and \code{space_distmat}.
\itemize{
\item \code{k_matrix}: a kernel matrix computed from the spatial distance matrix.
\item \code{beta}: a numeric parameter controlling the diffusion rate passed on to \code{\link[=k_diffusion]{k_diffusion()}}.
\item \code{space_distmat}: a numeric spatial distance matrix.
}}
}
\value{
a numeric vector of spatial statistics for a given cluster.
}
\description{
This function computes and returns the statistic of the spatial cost function for a given cluster.
}
