% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{convert_to_pop}
\alias{convert_to_pop}
\title{Convert a list of blob objects to a pop object}
\usage{
convert_to_pop(blob_list)
}
\arguments{
\item{blob_list}{a list of output from \code{\link[=blob_search]{blob_search()}}}
}
\value{
a \code{pop} object includes a list of the following objects.
\itemize{
\item \code{clust}: a numeric matrix of cluster assignments. Each row is a solution.
\item \code{summary}: a data frame of summary statistics.
\item \code{trace}: a data frame of summary statistics for tracing.
\item \code{n_filtered}: a data frame of numbers of filtered solutions.
\item \code{space_kmat_optim_out}: a list of \code{\link[stats:optim]{stats::optim()}} output from the optimisation of \eqn{\beta} in \code{\link[=distmat_to_kmat]{distmat_to_kmat()}}.
}
}
\description{
This function convert a list of blob objects to a pop object with combined output
from \code{\link[=blob_search]{blob_search()}} and counts of filtered solutions by category.
}
