% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{distmat_to_kmat}
\alias{distmat_to_kmat}
\title{Convert distance matrix to kernel matrix using diffusion kernel}
\usage{
distmat_to_kmat(distmat, k, w_knn = 7, l_normalise = TRUE, beta_par = 10)
}
\arguments{
\item{distmat}{a distance matrix.}

\item{k}{an integer of the number of clusters.}

\item{w_knn}{an integer of the \eqn{k^{th}} nearest neighbour used in local scaling for computing the adjacency matrix, passed onto \code{\link[=compute_adjacency]{compute_adjacency()}}. Default is 7.}

\item{l_normalise}{a Boolean to normalise the graph Laplacian, passed onto \code{\link[=compute_laplacian]{compute_laplacian()}}. Default is TRUE.}

\item{beta_par}{an numeric initial value for the parameter \eqn{\beta} to be optimised over for the diffusion kernel, passed onto \code{\link[=optim_beta]{optim_beta()}}. The range is \eqn{[0.001, \inf]}. Default is 10.}
}
\value{
a list of the following objects.
\itemize{
\item \code{kmat}: a diffusion kernel matrix.
\item \code{optim_out}: an output of \code{\link[stats:optim]{stats::optim()}}.
}
}
\description{
This function converts a distance matrix to a kernel matrix using a diffusion kernel.
}
\details{
This function takes a distance matrix as input and computes a weighted adjacency matrix using local scaling as proposed by Zelnik-Manor and Perona (2004).
A graph is constructed from the weighted adjacency matrix, represented by a graph Laplacian following Smola and Kondor (2003) and Chung (1997).
Then, it uses a diffusion kernel over the graph with optimised diffusion rate \eqn{\beta} to capture the global structure for kernel-based clustering (Kondor and Lafferty, 2002; Kondor and Vert, 2004).
\eqn{\beta} is optimised by minimising the absolute difference between the number of clusters and the effective rank of the kernel matrix where the effective rank,
which is a measure of effective dimensionality of a matrix, follows the definition in Roy and Vetterli (2007).
}
\references{
Chung, F. R. (1997). Spectral graph theory (Vol. 92). American Mathematical Soc..

Kondor, R. I., & Lafferty, J. (2002, July). Diffusion kernels on graphs and other discrete structures. In Proceedings of the 19th international conference on machine learning (Vol. 2002, pp. 315-322).

Kondor, R., & Vert, J. P. (2004). Diffusion kernels. kernel methods in computational biology, 171-192.

Roy, O., & Vetterli, M. (2007, September). The effective rank: A measure of effective dimensionality. In 2007 15th European signal processing conference (pp. 606-610). IEEE.

Smola, A. J., & Kondor, R. (2003, August). Kernels and regularization on graphs. In Learning theory and kernel machines: 16th annual conference on learning theory and 7th kernel workshop, COLT/kernel 2003, Washington, DC, USA, august 24-27, 2003. Proceedings (pp. 144-158). Berlin, Heidelberg: Springer Berlin Heidelberg.
}
\seealso{
\code{\link[=compute_adjacency]{compute_adjacency()}}, \code{\link[=compute_laplacian]{compute_laplacian()}}, \code{\link[=diffusion_kernel]{diffusion_kernel()}}, \code{\link[=optim_beta]{optim_beta()}}, \code{\link[=compute_erank]{compute_erank()}}
}
