% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{eval_blobs}
\alias{eval_blobs}
\title{Clustering evaluation}
\usage{
eval_blobs(data, space_distmat, sigma, crs = 4326)
}
\arguments{
\item{data}{a data frame.}

\item{space_distmat}{a numeric spatial distance matrix. Default is NULL.}

\item{sigma}{a numeric value of sigma passed on to \code{\link[=rbf]{rbf()}}. It must be specified when \code{space_distmat} is supplied.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}. Default is 4326.}

\item{space_distmethod}{a string of method used for distance computation. It must be one of "geodesic" or "euclidean". Default is "geodesic".}
}
\value{
a list of the following objects.
\itemize{
\item \code{summary}: a data frame of summary statistics.
\item \code{clust_below_size}: a numeric vector of clusters below the critical size.
}
}
\description{
This function evaluates clustering performance and returns a summary and clusters that are below the critical size.
}
\details{
When \code{space_distmat} is specified. Distance matrix implementation will ignore \code{space_distmethod}.

The critical size of a cluster is defined as \eqn{\frac{N}{2K}} where \eqn{N} is the number of data points and \eqn{k} is the number of clusters.
}
