% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{eval_blobs}
\alias{eval_blobs}
\title{Evaluate blobs}
\usage{
eval_blobs(
  data,
  coords = c(1, 2),
  age = 3,
  crs = 4326,
  space_clustmethod = "kmedoids",
  hull_convex_ratio = 0,
  space_kmat = NULL,
  space_distmat = NULL,
  space_distmethod = NULL,
  w_knn = NULL,
  l_normalise = NULL,
  beta_par = NULL,
  weights = NULL
)
}
\arguments{
\item{data}{a numeric matrix or data frame.}

\item{coords}{a vector of strings or numeric values indicating the columns of coordinates (longitude, latitide). Default is the first two columns.}

\item{age}{a string or numeric value indicating the column of age. Default is the third column.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}. Default is 4326.}

\item{space_clustmethod}{a string of the clustering method. Either "kmedoids" or "kkmeans" is available. Default is "kmedoids".}

\item{hull_convex_ratio}{a numeric value controlling the convexity of the hulls passed onto \code{\link[sf:geos_unary]{sf::st_concave_hull()}}. 1 returns convex and 0 maximally concave hulls. Default is 0.5.}

\item{space_kmat}{a kernel matrix computed from the spatial distance matrix. Default is NULL.}

\item{space_distmat}{a spatial distance matrix. Default is NULL.}

\item{space_distmethod}{a string of the method, used when \code{space_kmat} and \code{space_distmat} are not specified. Either "geodesic" or "euclidean" is available. Default is "geodesic".}

\item{w_knn}{an integer of the k-th nearest neighbour, used in local scaling for computing the adjacency matrix with \code{\link[=compute_adjacency]{compute_adjacency()}} when \code{space_kmat} is not supplied. Default is 7.}

\item{l_normalise}{a Boolean to normalise the graph Laplacian, passed onto \code{\link[=compute_laplacian]{compute_laplacian()}} when \code{space_kmat} is not supplied. Default is TRUE.}

\item{beta_par}{an numeric initial value for the parameter \eqn{\beta} to be optimised over for the diffusion kernel with \code{\link[=optim_beta]{optim_beta()}} when \code{space_kmat} is not supplied. The range is \eqn{[0.001, \inf]}. Default is 10.}

\item{weights}{a numeric vector of weights for each data point. Default is NULL.}
}
\value{
a list of the following objects.
\itemize{
\item \code{summary}: a data frame of summary statistics.
\item \code{clust_below_size}: a numeric vector of clusters below the critical size.
\item \code{space_kmat_optim_out}: an output of \code{\link[stats:optim]{stats::optim()}} from the optimisation of \eqn{\beta} when \code{space_kmat} is not supplied.
}
}
\description{
This function evaluates the performance of clustering.
}
\details{
The critical size of a cluster is defined as \eqn{\frac{n}{2k}} where \eqn{n} is the number of data points and \eqn{k} is the number of clusters.
}
