% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{eval_moo}
\alias{eval_moo}
\title{Evaluate multi-objective optimisation performance}
\usage{
eval_moo(pop_pareto_objspace_list)
}
\arguments{
\item{pop_pareto_objspace_list}{a list of data frames of the objective space from each batch.}
}
\value{
a data frame of MOO quality indicators.
}
\description{
This function evaluates multi-objective optimisation (MOO) performance.
}
\details{
The quality indicators include IGD, IGD+ and hypervolume commonly used in MOO.
They are computed using \code{\link[moocore:igd]{moocore::igd()}}, \code{\link[moocore:igd]{moocore::igd_plus()}}, and \code{\link[moocore:hypervolume]{moocore::hypervolume()}}. See \link[moocore:moocore-package]{moocore::moocore} for more information.

The lower and upper bounds for normalisation are extracted from the Pareto front considering all batches.
}
\seealso{
\link[moocore:moocore-package]{moocore::moocore}
}
