% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{find_blobs}
\alias{find_blobs}
\title{Assign clusters}
\usage{
find_blobs(data, k, r, space_distmat, sigma)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters.}

\item{r}{an numeric value of spatial relative weight. It must be \eqn{[0,1]}.}

\item{space_distmat}{a numeric spatial distance matrix.}

\item{sigma}{a numeric value of sigma passed on to \code{\link[=rbf]{rbf()}}.}
}
\value{
a data frame with assigned clusters as a column.
}
\description{
This function assigns clusters for a given k and r.
}
\details{
Gaussian kernel is applied to compute cluster centroids in space.
}
