% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{find_blobs}
\alias{find_blobs}
\title{Assign clusters}
\usage{
find_blobs(data, k, r, space_kmat, age = 3)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters.}

\item{r}{an numeric value of spatial relative weight. It must be \eqn{[0,1]}.}

\item{space_kmat}{a kernel matrix computed from the spatial distance matrix.}

\item{age}{a string or numeric value indicating the column of age. Default is the third column.}
}
\value{
a data frame with assigned clusters as a column.
}
\description{
This function assigns clusters for a given k and r.
}
\details{
A diffusion kernel is applied to compute the distance to the centroid in space.
}
\references{
Dhillon, I. S., Guan, Y., & Kulis, B. (2004, August 22). Kernel k-means: spectral clustering and normalized cuts. Proceedings of the Tenth ACM SIGKDD International Conference on Knowledge Discovery and Data Mining. KDD04: ACM SIGKDD International Conference on Knowledge Discovery and Data Mining, Seattle WA USA. https://doi.org/10.1145/1014052.1014118

Kondor, R. I., & Lafferty, J. (2002, July). Diffusion kernels on graphs and other discrete structures. In Proceedings of the 19th international conference on machine learning (Vol. 2002, pp. 315-322).

Kondor, R., & Vert, J. P. (2004). Diffusion kernels. kernel methods in computational biology, 171-192.

Schölkopf, B., Smola, A., & Müller, K.-R. (1998). Nonlinear component analysis as a kernel eigenvalue problem. Neural Computation, 10(5), 1299–1319.
}
\seealso{
\code{\link[=compute_spacecost]{compute_spacecost()}}
}
