% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_simdata.R
\name{gen_circle_coords}
\alias{gen_circle_coords}
\title{Generate coordinates on a circle outline}
\usage{
gen_circle_coords(radius, n)
}
\arguments{
\item{radius}{a numeric value of radius from the center.}

\item{n}{an integer of the number of vertices.}
}
\value{
a numeric matrix of coordinates.
}
\description{
This function generates coordinates of equal distances on a circle outline centered at (0,0).
}
