% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{intersects_bool}
\alias{intersects_bool}
\title{Check for spatial intersects}
\usage{
intersects_bool(
  data,
  clust = NULL,
  coords = c(1, 2),
  crs = 4326,
  hull_convex_ratio = 0
)
}
\arguments{
\item{data}{a numeric matrix or data frame.}

\item{clust}{a numeric vector of the cluster assignment. Default is NULL.}

\item{coords}{a vector of strings or numeric values indicating the columns of coordinates (longitude, latitide). Default is the first two columns.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}}. Default is 4326.}

\item{hull_convex_ratio}{a numeric value controlling the convexity of the hulls passed onto \code{\link[sf:geos_unary]{sf::st_concave_hull()}}. 1 returns convex and 0 maximally concave hulls. Default is 0.}
}
\value{
TRUE or FALSE
}
\description{
This function checks and returns a Boolean if any clusters intersect in space.
}
