% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_space}
\alias{plot_space}
\title{Plot data in space}
\usage{
plot_space(
  data,
  clust = NA,
  space = "earth",
  coords = c(1, 2),
  crs = NULL,
  size = 1,
  alpha = 0.8,
  hull = FALSE,
  hull_convex_ratio = 0.8,
  weights = NULL
)
}
\arguments{
\item{data}{a data frame}

\item{clust}{a numeric vector of cluster assignment.}

\item{space}{a string of the space data plotted in. It must be either "earth" or "euclidean".}

\item{coords}{a vector of strings or numeric values indicating the columns of coordinates (longitude, latitide). Default is the first two columns.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}} and \code{\link[sf:st_transform]{sf::st_transform()}}. Default is NULL.}

\item{size}{size for \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}

\item{alpha}{See <\code{\link[ggplot2:aes_colour_fill_alpha]{aes-colour-fill-alpha}}>.}

\item{hull}{a Boolean. Should convex hulls be drawn? Default is FALSE.}

\item{hull_convex_ratio}{a numeric value controlling the convexity of the hulls passed onto \code{\link[sf:geos_unary]{sf::st_concave_hull()}}. 1 returns convex and 0 maximally concave hulls. Default is 0.5.}

\item{weights}{a numeric vector of weights for each data point, used to indicate them by data point colours. Default is NULL.}
}
\value{
a \code{ggplot} object.
}
\description{
This function plots data in space.
}
\seealso{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}}, \code{\link[sf:st_transform]{sf::st_transform()}}
}
