% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_space}
\alias{plot_space}
\title{Plot data in space}
\usage{
plot_space(
  data,
  clust = NA,
  space = "earth",
  hull = F,
  crs = 3035,
  buffer = 5e+05,
  lab = NULL,
  palette = NULL
)
}
\arguments{
\item{data}{a data frame}

\item{clust}{a numeric vector of cluster assignemnt. Default is NA.}

\item{space}{a string of the space data plotted in. It must be either "earth" or "euclidean". Default is "earth".}

\item{hull}{a logical operator. Should convex hulls be drawn? Default is F.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}} and \code{\link[sf:st_transform]{sf::st_transform()}}. Default is 3035.}

\item{buffer}{a numeric value of the buffer between the data points and edges of the plot. Default is 500000 for \code{crs = 3035}.}

\item{lab}{a string vector of length 2. The first one is passed on to \code{\link[ggplot2:labs]{ggplot2::xlab()}} and the second \code{\link[ggplot2:labs]{ggplot2::ylab()}}. Default is NULL.}

\item{palette}{See \code{value} of \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}. Default is NULL. This uses the palette presets from \code{\link[MetBrewer:met.brewer]{MetBrewer::met.brewer()}}.}
}
\value{
a ggplot.
}
\description{
This function plots data in space.
}
\seealso{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}}, \code{\link[sf:st_transform]{sf::st_transform()}}, \code{\link[ggplot2:labs]{ggplot2::labs()}}, \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}, \code{\link[MetBrewer:met.brewer]{MetBrewer::met.brewer()}}
}
