% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_space}
\alias{plot_space}
\title{Plot data in space}
\usage{
plot_space(
  data,
  clust,
  space = c("earth", "euclidean"),
  coords = c(1, 2),
  crs = NULL,
  hull = FALSE,
  hull_convex_ratio = 0.8
)
}
\arguments{
\item{data}{a data frame}

\item{clust}{a numeric vector of cluster assignment.}

\item{space}{a string of the space data plotted in. It must be either "earth" or "euclidean".}

\item{coords}{a vector of strings or numeric values indicating the columns of coordinates (longitude, latitide). Default is the first two columns.}

\item{crs}{a numeric value of the Coordinate Reference System passed on to \code{\link[sf:st_as_sf]{sf::st_as_sf()}} and \code{\link[sf:st_transform]{sf::st_transform()}}. Default is NULL.}

\item{hull}{a Boolean. Should convex hulls be drawn? Default is FALSE.}

\item{hull_convex_ratio}{a numeric value controlling the convexity of the hulls passed onto \code{\link[sf:geos_unary]{sf::st_concave_hull()}}. 1 returns convex and 0 maximally concave hulls. Default is 0.}
}
\value{
a \code{ggplot} object.
}
\description{
This function plots data in space.
}
\seealso{
\code{\link[sf:st_as_sf]{sf::st_as_sf()}}, \code{\link[sf:st_transform]{sf::st_transform()}}
}
