% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plotting.R
\name{plot_time}
\alias{plot_time}
\title{Plot data in time}
\usage{
plot_time(data, clust, orientation = "portrait", lab = NULL, palette = NULL)
}
\arguments{
\item{data}{a data frame}

\item{clust}{a numeric vector of cluster assignemnt.}

\item{orientation}{a string of orientation of the plot. It must be either "portrait" or "landscape".}

\item{lab}{a string of label passed on to \code{\link[ggplot2:labs]{ggplot2::xlab()}}. Default is NULL.}

\item{palette}{See \code{value} of \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}. Default is NULL. This uses the palette presets from \code{\link[MetBrewer:met.brewer]{MetBrewer::met.brewer()}}.}
}
\value{
a ggplot.
}
\description{
This function plots data in time
}
\seealso{
\code{\link[ggplot2:labs]{ggplot2::labs()}}, \code{\link[ggplot2:scale_manual]{ggplot2::scale_colour_manual()}}, \code{\link[MetBrewer:met.brewer]{MetBrewer::met.brewer()}}
}
