% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{rbf}
\alias{rbf}
\title{Radial basis function}
\usage{
rbf(distance, sigma)
}
\arguments{
\item{distance}{a numeric or complex vector of distances.}

\item{sigma}{a numeric value of sigma.}
}
\value{
a numeric or complex vector of distances \eqn{[0,1} in the feature space.
}
\description{
This function kernelises the distance using the radial basis function. It is also known as the Gaussian kernel.
}
