% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{reorder_clust}
\alias{reorder_clust}
\title{Reorder clusters}
\usage{
reorder_clust(clust)
}
\arguments{
\item{clust}{a numeric vector of cluster assignment.}
}
\value{
a numeric vector of reordered cluster assignment.
}
\description{
This function reorders and returns a vector of cluster assignment in ascending order.
}
