% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_localsearch.R
\name{start_blobs}
\alias{start_blobs}
\title{Assign the starting cluster members}
\usage{
start_blobs(data, k, space_mat, method = "kmedoids", random_start = FALSE)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters.}

\item{space_mat}{a spatial distance matrix or a kernel matrix computed from the spatial distance matrix.}

\item{method}{a string of the clustering method. Either "kmedoids" or "kkmeans" is available. Default is "kmedoids".}

\item{random_start}{a Boolean to randomise starting cluster members. Default is FALSE.}
}
\value{
a data frame with assigned starting clusters as a column.
}
\description{
This function assigns the starting cluster members by approximating the maximum spatial separation between points or by random assignment.
}
