% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_core.R
\name{start_blobs}
\alias{start_blobs}
\title{Assign the starting cluster members by approximating the maximum spatial separation}
\usage{
start_blobs(data, k, space_distmat)
}
\arguments{
\item{data}{a data matrix or data frame.}

\item{k}{an integer of the number of clusters.}

\item{space_distmat}{a numeric spatial distance matrix.}
}
\value{
a data frame with assigned starting clusters as a column.
}
\description{
Assign the starting cluster members by approximating the maximum spatial separation between points.
}
